Move into assets folder
