+++ src/components/LeafletMap.jsx
@@ -9,7 +9,7 @@ import 'leaflet.markercluster/dist/MarkerCluster.css'
 import 'leaflet.markercluster/dist/MarkerCluster.Default.css'
 import 'leaflet/dist/leaflet.css'
 import '../styles/leafletMapContainerStyles.css'
-import customIcon from '../styles/Icons/map-pin.png'
+import customIcon from './assets/map-pin.png'
 import usePrevious from '../hooks/usePrevious'
 import theme from '../styles/theme'
 import MapAndTableControls from './MapAndTableControls/MapAndTableControls'
