+++ src/components/MapAndTableControls/MapAndTableControls.jsx
@@ -5,8 +5,8 @@ import FilterIndicatorPill from './FilterIndicatorPill'
 import ViewToggle from './ViewToggle'
 import { mediaQueryTabletLandscapeOnly } from '../../styles/mediaQueries'
 import { ButtonSecondary } from '../generic'
-import zoomToSelectedSites from '../styles/Icons/zoom_to_selected_sites.svg'
-import zoomToFiltered from '../styles/Icons/zoom_to_filtered.svg'
+import zoomToSelectedSites from '../assets/zoom_to_selected_sites.svg'
+import zoomToFiltered from '../assets/zoom_to_filtered.svg'
 import useResponsive from '../../hooks/useResponsive'
 import { useFilterProjectsContext } from '../../context/FilterProjectsContext'
 import { URL_PARAMS } from '../../constants/constants'
