+++ src/components/MapAndTableControls/ViewToggle.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useRef } from 'react'
 import styled from 'styled-components'
 import { ButtonPrimary, ButtonSecondary } from '../generic'
-import { IconMapOutline, IconTable, IconTrayDownload } from '../../styles/Icons/dashboardOnlyIcons'
+import { IconMapOutline, IconTable, IconTrayDownload } from '../assets/dashboardOnlyIcons'
 import { useLocation, useNavigate } from 'react-router-dom'
 import { formatProjectDataHelper } from '../../helperFunctions'
 import { CSVLink } from 'react-csv'
